refactor: break down monolithic components into modular architecture
- Decompose AboutSection, ContactSection, PortfolioSection, and ServicesSection into smaller, focused components
- Create organized component folders (about/, contact/, portfolio/, services/) with index exports
- Extract reusable components like AnimatedCounter, SocialLinks, ContactForm, etc.
- Replace placeholder portfolio images with actual project showcases (1.webp, 2.webp)
- Update portfolio data with real project information (Fitness Academy, Gaming Gift Store)
- Optimize portfolio images using webp format for better performance
- Improve component maintainability and reusability through better separation of concerns
