+++ src/components/about/index.js
@@ -1,0 +1,7 @@
+export { default as AboutBadge } from './AboutBadge';
+export { default as AboutHeading } from './AboutHeading';
+export { default as AboutImage } from './AboutImage';
+export { default as AboutContent } from './AboutContent';
+export { default as AboutStats } from './AboutStats';
+export { default as AboutLogos } from './AboutLogos';
+export { default as AnimatedCounter } from './AnimatedCounter';
