+++ src/components/contact/index.js
@@ -1,0 +1,7 @@
+export { default as ContactLabel } from './ContactLabel';
+export { default as ContactHeading } from './ContactHeading';
+export { default as ContactDescription } from './ContactDescription';
+export { default as ContactInfo } from './ContactInfo';
+export { default as SocialLinks } from './SocialLinks';
+export { default as ContactForm } from './ContactForm';
+export { default as ContactLeftSide } from './ContactLeftSide';
