+++ src/components/portfolio/PortfolioData.js
@@ -1,0 +1,34 @@
+// Import your images
+import project1 from "../../images/projects/1.webp";
+import project2 from "../../images/projects/2.webp";
+import project3 from "../../images/projects/Trailblaze Outfitters Case Study (1).avif";
+
+export const projects = [
+	{
+		title: "🏋️‍♂️ Fitness Academy & Gym Training Website",
+		description:
+			"A highly interactive and modern web platform for a gym + fitness training academy. Built using GatsbyJS, styled with TailwindCSS, and animated via Framer Motion, this site combines performance, sleek design, and engaging user experience.",
+    	description2:
+			"",
+		image: project1,
+	},
+	{
+		title: "🕹️ Fullstack Project: Gaming Gift Store & Merch Platform",
+		description:
+			"A fully responsive and highly interactive eCommerce web application where users can purchase gift packs for games, T-shirts, figurines, and more. Integrated real-time wallet, vendor APIs, and animations for a seamless shopping experience.",
+    description2:"",
+		image: project2,
+	},
+	{
+		title: "Artisan Coffee",
+		description:
+			"Trailblaze Outfitters wanted their website to inspire adventure seekers.",
+    description2:
+    "We created an energetic, bold digital experience featuring striking imagery, interactive product displays, and a seamless e-commerce platform. The final result is a website as rugged and reliable as the gear they offer.",
+		image: project3,
+	},
+];
+
+const PortfolioData = () => null; // This is just a data export component
+
+export default PortfolioData;
