+++ src/components/portfolio/index.js
@@ -1,0 +1,6 @@
+export { default as PortfolioData, projects } from './PortfolioData';
+export { default as PortfolioImage } from './PortfolioImage';
+export { default as PortfolioHeader } from './PortfolioHeader';
+export { default as PortfolioContent } from './PortfolioContent';
+export { default as PortfolioNavigation } from './PortfolioNavigation';
+export { default as PortfolioGradients } from './PortfolioGradients';
