+++ src/components/services/index.js
@@ -1,0 +1,5 @@
+export { default as ServicesLabel } from './ServicesLabel';
+export { default as ServicesHeading } from './ServicesHeading';
+export { default as ServicesList } from './ServicesList';
+export { default as ServicesDescription } from './ServicesDescription';
+export { default as ServicesImage } from './ServicesImage';
